# server/internal/handlers/auth.go
package handlers

import (
	"encoding/json"
	"errors"
	"log"
	"net/http"
	"time"

	"github.com/Harshitttttttt/Swayamsevak/server/internal/auth"
	"github.com/Harshitttttttt/Swayamsevak/server/internal/handlers/dto"
)

// AuthHandler contains HTTP handlers for authentication
type AuthHandler struct {
	authService     *auth.AuthService
	refreshTokenTTL time.Duration
	cookieSecure    bool
}

// NewAuthHandler creates a new Auth handler
func NewAuthHandler(authService *auth.AuthService, refreshTTL time.Duration, cookieSecure bool) *AuthHandler {
	return &AuthHandler{
		authService:     authService,
		refreshTokenTTL: refreshTTL,
		cookieSecure:    cookieSecure,
	}
}

const refreshTokenCookieName = "refresh_token"

func (h *AuthHandler) setRefreshTokenCookie(w http.ResponseWriter, tokenString string) {
	http.SetCookie(w, &http.Cookie{
		Name:     refreshTokenCookieName,
		Value:    tokenString,
		HttpOnly: true,
		Secure:   h.cookieSecure,
		SameSite: http.SameSiteLaxMode,
		Path:     "/api/",
		MaxAge:   int(h.refreshTokenTTL.Seconds()),
	})
}

func (h *AuthHandler) clearRefreshTokenCookie(w http.ResponseWriter) {
	http.SetCookie(w, &http.Cookie{
		Name:     refreshTokenCookieName,
		Value:    "",
		HttpOnly: true,
		Expires:  time.Unix(0, 0),
		Path:     "/api",
	})
}

// Register godoc
// @Summary      Register a new user
// @Description  Create a new user account with email, username, and password
// @Tags         Authentication
// @Accept       json
// @Produce      json
// @Param        request body dto.RegisterRequest true "Registration credentials"
// @Success      201 {object} dto.RegisterResponse
// @Failure      400 {string} string "Invalid Request Payload"
// @Failure      409 {string} string "Email already in use"
// @Failure      500 {string} string "Error Creating User"
// @Router       /auth/register [post]
func (h *AuthHandler) Register(w http.ResponseWriter, r *http.Request) {
	// Parse the request body
	var req dto.RegisterRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid Request Payload", http.StatusBadRequest)
		return
	}

	// Validate Input
	if req.Email == "" || req.Username == "" || req.Password == "" {
		http.Error(w, "Email, Username and Password are required", http.StatusBadRequest)
		return
	}

	// Call the AuthService to register the user
	user, err := h.authService.Register(req.Email, req.Username, req.Password)
	if err != nil {
		if errors.Is(err, auth.ErrEmailInUse) {
			http.Error(w, "Email already in use", http.StatusConflict)
			return
		}

		// Log the underlying error for debugging
		log.Printf("register: failed to create user: %v", err)
		http.Error(w, "Error Creating User", http.StatusInternalServerError)
		return
	}

	// Return the created user (without sensitive data)
	response := dto.RegisterResponse{
		ID:       user.ID.String(),
		Email:    user.Email,
		Username: user.Username,
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(response)
}

// Login godoc
// @Summary      Login user
// @Description  Authenticate user and receive JWT access token
// @Tags         Authentication
// @Accept       json
// @Produce      json
// @Param        credentials body dto.LoginRequest true "Login credentials"
// @Success      200 {object} dto.LoginResponse "Access token; refresh token is set as an HttpOnly cookie"
// @Failure      400 {string} string "Invalid Request Payload"
// @Failure      401 {string} string "Invalid Credentials"
// @Failure      500 {string} string "Internal Server Error"
// @Router       /auth/login [post]
func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {

	// Parse the request body
	var req dto.LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid Request Payload", http.StatusBadRequest)
		return
	}

	// Attempt to login
	accessToken, refreshToken, err := h.authService.LoginWithRefresh(req.Email, req.Password, h.refreshTokenTTL)
	if err != nil {
		if errors.Is(err, auth.ErrInvalidCredentials) {
			http.Error(w, "Invalid Credentials", http.StatusUnauthorized)
		} else {
			log.Printf("Internal Server Error: %v", err)
			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
		}
		return
	}

	// Set the refresh token as an HttpOnly cookie
	h.setRefreshTokenCookie(w, refreshToken)

	// Return the token
	response := dto.LoginResponse{
		AccessToken: accessToken,
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// RefreshToken godoc
// @Summary      Refresh access token
// @Description  Get a new access token using a valid refresh token
// @Tags         Authentication
// @Accept       json
// @Produce      json
// @Param        token body dto.RefreshRequest false "Refresh token (optional when cookie is used)"
// @Success      200 {object} dto.RefreshResponse
// @Failure      400 {string} string "Invalid Request Payload"
// @Failure      401 {string} string "Invalid or Expired refresh token"
// @Failure      500 {string} string "Internal Server Error"
// @Router       /auth/refresh [post]
func (h *AuthHandler) RefreshToken(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	// Parse the request body
	var req dto.RefreshRequest
	_ = json.NewDecoder(r.Body).Decode(&req) // We ignore error here because we are reading token from cookie

	// Get the refresh token from cookie
	var tokenStr string
	cookie, err := r.Cookie(refreshTokenCookieName)
	if err == nil {
		tokenStr = cookie.Value
	} else {
		tokenStr = req.RefreshToken
	}

	if tokenStr == "" {
		http.Error(w, "Missing Refresh Token", http.StatusUnauthorized)
		return
	}

	// Attempt to refresh the token
	accessToken, newRefreshToken, err := h.authService.RefreshAccessToken(tokenStr, h.refreshTokenTTL)
	if err != nil {
		if errors.Is(err, auth.ErrInvalidToken) || errors.Is(err, auth.ErrExpiredToken) {
			http.Error(w, "Invalid or Expired refresh token", http.StatusUnauthorized)
		} else {
			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
		}
		return
	}

	// Set new refresh token cookie (rotate in place)
	h.setRefreshTokenCookie(w, newRefreshToken)

	// Return the new access token
	response := &dto.RefreshResponse{
		Token: accessToken,
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// Logout godoc
// @Summary      Logout user
// @Description  Revoke refresh token. The refresh token is read from the HttpOnly cookie named "refresh_token" by default; clients may optionally pass it in the request body.
// @Tags         Authentication
// @Accept       json
// @Produce      json
// @Param        token body dto.LogoutRequest false "Refresh token (optional when cookie is used)"
// @Success      204 "No Content"
// @Failure      400 {string} string "Invalid Request Payload"
// @Failure      401 {string} string "Invalid refresh token"
// @Router       /auth/logout [post]
func (h *AuthHandler) Logout(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	var req dto.LogoutRequest
	_ = json.NewDecoder(r.Body).Decode(&req) // We ignore error here because we are reading token from cookie

	tokenStr := req.RefreshToken
	if cookie, err := r.Cookie(refreshTokenCookieName); err == nil {
		tokenStr = cookie.Value
	}

	if tokenStr == "" {
		http.Error(w, "Invalid Request Payload", http.StatusBadRequest)
		return
	}

	if err := h.authService.Logout(tokenStr); err != nil {
		http.Error(w, "Invalid refresh token", http.StatusUnauthorized)
		return
	}

	// Clear cookie on logout
	h.clearRefreshTokenCookie(w)
	w.WriteHeader(http.StatusNoContent)
}